Testing weapon goals...
Renaming some categories to match in-game text.
Merging categories arbitrarily.
Using vanilla tables to roll loot (including 1-hand and 2-hand tables).
Removing daggers (don't exist).
Is due for a re-testing since most of the changes came after checking for in-game drops but it's safe enough to do that with the final testing rounds, which should be semi-extensive given the scope of the changes (re-tested tables with GUTS, they're all good and their names didn't change so everything **should** work as when first testing it, hopefully).
Some of the weapons rolled are internal with weird display names but if the engine doesn't handle that internally when using the vanilla tables, I'm gonna leave it as-is for now (the drops are actually normal items other than the names, with a single exception so far).