Drop 1-2 spells instead of always 1...
Not finding good spells for your build is a major break in pacing.
It's possible to get many useless spells in a row (ranged, ranged passives)...
Early on it's important to fill slots.
Later it's important to always evolve rather than not change spells for 10+ levels.
Later it's also possible to not get a higher rank upgrade for a long time...
As there aren't many ranks and they're spaced far apart, it's even worse.
The hard limit of 4 spells means the power ceiling isn't changed...
Just how fast you can reach the ceiling (potentially, not realistically).
A 150% increase in drop quantity is mininal...
Makes players excited to get a second drop but still content with 1?
Players who may want to focus on bosses/dungeons can fill their spells more easily.