Moving extra spells to 10% drop on champions...
While the extra spell is a great addition, it doesn't help for players struggling to finish a level.
Having a small drop chance on champions is more granular and accessible, fitting the purpose of the extra spell.
I wouldn't import the champion table for this only but since it's already here for map drops, might as well use it.